feat(gpu): signed scalar sub
