+++ tfhe/src/high_level_api/integers/unsigned/scalar_ops.rs
@@ -503,7 +503,7 @@ generic_integer_impl_scalar_operation!(
                  InternalServerKey::Cuda(cuda_key) => {
                     let inner_result = with_thread_local_cuda_stream(|stream| {
                         cuda_key.key.scalar_sub(
-                            &lhs.ciphertext.on_gpu(), rhs, stream
+                            &*lhs.ciphertext.on_gpu(), rhs, stream
                         )
                     });
                     RadixCiphertext::Cuda(inner_result)
