+++ tfhe/src/integer/server_key/radix_parallel/tests_unsigned/mod.rs
@@ -2,6 +2,7 @@ pub(crate) mod test_add;
 pub(crate) mod test_mul;
 pub(crate) mod test_neg;
 pub(crate) mod test_scalar_add;
+pub(crate) mod test_scalar_sub;
 pub(crate) mod test_sub;
 
 use super::tests_cases_unsigned::*;
@@ -597,9 +598,6 @@ create_parametrized_test!(integer_unchecked_scalar_rotate_left);
 create_parametrized_test!(integer_default_scalar_rotate_right);
 create_parametrized_test!(integer_default_scalar_rotate_left);
 create_parametrized_test!(integer_default_scalar_div_rem);
-create_parametrized_test!(integer_smart_scalar_sub);
-create_parametrized_test!(integer_default_scalar_sub);
-create_parametrized_test!(integer_default_overflowing_scalar_sub);
 create_parametrized_test!(integer_smart_if_then_else);
 create_parametrized_test!(integer_default_if_then_else);
 create_parametrized_test!(integer_trim_radix_msb_blocks_handles_dirty_inputs);
@@ -982,14 +980,6 @@ where
 // Smart Scalar Tests
 //=============================================================================
 
-fn integer_smart_scalar_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::smart_scalar_sub_parallelized);
-    smart_scalar_sub_test(param, executor);
-}
-
 fn integer_smart_small_scalar_mul<P>(param: P)
 where
     P: Into<PBSParameters>,
@@ -1133,23 +1123,6 @@ where
 //=============================================================================
 // Default Scalar Tests
 //=============================================================================
-
-fn integer_default_scalar_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::scalar_sub_parallelized);
-    default_scalar_sub_test(param, executor);
-}
-
-fn integer_default_overflowing_scalar_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor =
-        CpuFunctionExecutor::new(&ServerKey::unsigned_overflowing_scalar_sub_parallelized);
-    default_overflowing_scalar_sub_test(param, executor);
-}
 
 fn integer_default_scalar_bitand<P>(param: P)
 where
